test: separate tests by app name
